feat(license): integrate LicensePromotion component and clean up related code
