+++ src/components.d.ts
@@ -37,6 +37,7 @@ declare module 'vue' {
     InputWithPlaceholderSelect: typeof import('./components/InputWithPlaceholderSelect.vue')['default']
     InputWithUnit: typeof import('./components/InputWithUnit.vue')['default']
     KeyAndValueEditor: typeof import('./components/KeyAndValueEditor.vue')['default']
+    LicensePromotion: typeof import('./components/LicensePromotion.vue')['default']
     LinkButton: typeof import('./components/Buttons/LinkButton.vue')['default']
     ListenerDrawer: typeof import('./components/ListenerDrawer/ListenerDrawer.vue')['default']
     MarkdownContent: typeof import('./components/MarkdownContent.vue')['default']
